+++ src/contracts/createTx/types.ts
@@ -1,7 +1,7 @@
 import { SettingsControllerUpdate } from "types/contracts";
 import { AccountType, ERC20Deposit, NewAST } from "types/contracts/evm";
 import { Allowance, Transfer } from "types/contracts/evm/erc20";
-import { NewTransaction } from "types/contracts/evm/multisig";
+import { NewTransaction } from "types/contracts/multisig";
 import { Tupleable } from "types/evm";
 import { Contract } from "types/lists";
