+++ src/pages/Admin/Charity/Dashboard/Table.tsx
@@ -1,5 +1,5 @@
 import { Link } from "react-router-dom";
-import { Transaction } from "types/contracts/evm/multisig";
+import { Transaction } from "types/contracts/multisig";
 import Icon from "components/Icon";
 import TableSection, { Cells } from "components/TableSection";
 import { adminRoutes } from "constants/routes";
