--- src/types/contracts/cw3.ts
@@ -1,6 +1,0 @@
-import { TransactionStatus } from "types/lists";
-
-export type PageOptions = {
-  range: [number, number];
-  status: TransactionStatus;
-};
