+++ src/types/contracts/index.ts
@@ -1,4 +1,3 @@
-export * from "./cw3";
 export * from "./account";
 export * from "./common";
 export * from "./cw20";
